+++ src/components/Tabs.tsx
@@ -1,29 +1,14 @@
 import classNames from 'classnames';
-import { Link, useParams } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { Tab } from '../types/Tab';
 
-const tabs: Tab[] = [
-  {
-    id: 'tab-1',
-    title: 'Tab 1',
-    content: 'Some text 1',
-  },
-  {
-    id: 'tab-2',
-    title: 'Tab 2',
-    content: 'Some text 2',
-  },
-  {
-    id: 'tab-3',
-    title: 'Tab 3',
-    content: 'Some text 3',
-  },
-];
+type Props = {
+  tabs: Tab[];
+  selectedTabId?: string;
+};
 
-export const Tabs = () => {
-  const { tabId } = useParams();
-
-  const selectedTab = tabs.find(tab => tab.id === tabId);
+export const Tabs: React.FC<Props> = ({ tabs, selectedTabId }) => {
+  const selectedTab = tabs.find(tab => tab.id === selectedTabId);
 
   return (
     <>
@@ -34,24 +19,17 @@ export const Tabs = () => {
               key={tab.id}
               data-cy="Tab"
               className={classNames({
-                'is-active': tab.id === tabId,
+                'is-active': tab.id === selectedTabId,
               })}
             >
-              <Link to={`/tabs/${tab.id}`}>
-                {tab.title}
-              </Link>
+              <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
             </li>
           ))}
         </ul>
       </div>
 
-      <div
-        className="block"
-        data-cy="TabContent"
-      >
-        {selectedTab
-          ? selectedTab.content
-          : 'Please select a tab'}
+      <div className="block" data-cy="TabContent">
+        {selectedTab ? selectedTab.content : 'Please select a tab'}
       </div>
     </>
   );
